refactor(locality): extract getOwnAddresses/isLocallyOrchestratable to servers/shared/locality.js
The providers reconciler (owner-asserts sync design, next commits) needs the
F-INSTALL-10 physical-locality predicate from servers/shared, which must not
import from servers/gateway. gpu-orchestrator re-exports both functions so
its existing importers and tests are unchanged.
