+++ servers/gateway/gpu-orchestrator.js
@@ -42,8 +42,8 @@ import { spawn } from "node:child_process";
 import { existsSync } from "node:fs";
 import { fileURLToPath } from "node:url";
 import { join, dirname, resolve } from "node:path";
-import { networkInterfaces } from "node:os";
 import { loadProviders as loadCachedProviders } from "../shared/providers.js";
+import { getOwnAddresses, isLocallyOrchestratable } from "../shared/locality.js";
 import {
   setResidencyInitialized, recordResidency, releaseResidency,
   pruneResidency, getProviderHealth,
@@ -56,46 +56,10 @@ function loadProviders() {
   return loadCachedProviders();
 }
 
-/**
- * F-INSTALL-10 — physical locality gate.
- *
- * The orchestrator's job is `docker compose up/stop` on THIS machine, so the
- * only trustworthy signal is whether the provider's baseUrl points AT this
- * machine (loopback or one of our own interface addresses). The providers
- * `host` column cannot be used: it syncs fleet-wide with the seeding
- * instance's perspective baked in (live fleet: grackle's own embed row says
- * host='grackle-5fc01ac74463b6f4', crow's bundles say 'local' everywhere),
- * so a host-string gate either breaks a peer keeping its own bundle resident
- * or lets a fresh install start the maintainer-lab's bundles.
- */
-// Bridge/virtual interfaces carry SHARED-SUBNET gateway IPs (every docker
-// host has 172.17.0.1; libvirt ships 192.168.122.1) — never machine identity
-// (R2-M1). Skip them so a peer's hypothetical bridge-IP baseUrl can't
-// false-match here.
-const VIRTUAL_IF_RE = /^(docker|br-|veth|virbr|vmnet|lxc|cni)/;
-
-export function getOwnAddresses() {
-  const own = new Set(["localhost", "127.0.0.1", "::1"]);
-  try {
-    for (const [ifname, addrs] of Object.entries(networkInterfaces())) {
-      if (VIRTUAL_IF_RE.test(ifname)) continue;
-      for (const a of addrs || []) own.add(a.address);
-    }
-  } catch {}
-  return own;
-}
-
-export function isLocallyOrchestratable(p, ownAddrs = getOwnAddresses()) {
-  if (!p?.baseUrl) return false;
-  try {
-    // WHATWG URL keeps brackets on IPv6 hostnames ("[::1]"); interface
-    // addresses don't have them.
-    const h = new URL(p.baseUrl).hostname.replace(/^\[|\]$/g, "");
-    return ownAddrs.has(h);
-  } catch {
-    return false;
-  }
-}
+// F-INSTALL-10 physical locality gate — implementation lives in
+// servers/shared/locality.js (shared with the providers reconciler's
+// owner-asserts gate). Re-exported here for existing importers/tests.
+export { getOwnAddresses, isLocallyOrchestratable };
 
 const READINESS_TIMEOUT_MS = 240_000;  // vLLM VL warm takes 2.5-3.5 min on 16 GB
 const READINESS_POLL_MS = 2_000;
